feat: Exception handling and BeanValidation

+++ src/main/java/com/devsuperior/challenge3/services/ClientService.java
@@ -3,6 +3,8 @@ package com.devsuperior.challenge3.services;
 import com.devsuperior.challenge3.dto.ClientDTO;
 import com.devsuperior.challenge3.entities.Client;
 import com.devsuperior.challenge3.repositories.ClientRepository;
+import com.devsuperior.challenge3.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,7 +19,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ClientDTO(client);
     }
 
@@ -36,14 +39,23 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyToEntity(entity, dto);
-        return new ClientDTO(repository.save(entity));
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyToEntity(entity, dto);
+            return new ClientDTO(repository.save(entity));
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
     @Transactional
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+            repository.deleteById(id);
     }
 
     private void copyToEntity(Client entity, ClientDTO dto) {
